migrated mod specific injection code into their own mixin classes
