+++ src/main/java/com/github/thebrochacho/dws/mixin/mixins/common/dws/galacticraft/ContainerPlayerMixin.java
@@ -1,0 +1,73 @@
+package com.github.thebrochacho.dws.mixin.mixins.common.dws.galacticraft;
+
+import com.github.thebrochacho.dws.interfaces.dws.IAddsGCSlots;
+import com.github.thebrochacho.dws.inventory.slots.SlotDWS;
+import com.github.thebrochacho.dws.util.SlotLayoutManager;
+import micdoodle8.mods.galacticraft.core.entities.player.GCPlayerStats;
+import micdoodle8.mods.galacticraft.core.proxy.ClientProxyCore;
+import micdoodle8.mods.galacticraft.core.util.PlayerUtil;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
+import net.minecraft.entity.player.InventoryPlayer;
+import net.minecraft.inventory.Container;
+import net.minecraft.inventory.ContainerPlayer;
+import net.minecraft.inventory.IInventory;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import static com.github.thebrochacho.dws.inventory.slots.SlotDWS.SlotType.*;
+import static com.github.thebrochacho.dws.inventory.slots.SlotDWS.SlotType.GC_OXYGEN_TANK;
+
+@Mixin(ContainerPlayer.class)
+public abstract class ContainerPlayerMixin extends Container implements IAddsGCSlots {
+
+    private IInventory galacticraftAccessories = null;
+
+    private static int GC_SLOT_START = -1;
+
+    @Override
+    public int getGCSlotStart() {
+        return GC_SLOT_START;
+    }
+
+    @Inject(method = "<init>",
+            at = @At("RETURN"),
+            require = 1)
+    private void injectGalacticraftSlots(InventoryPlayer inventoryPlayer, boolean isLocalWorld, EntityPlayer player, CallbackInfo ci) {
+        if (GC_SLOT_START == -1) {
+            GC_SLOT_START = this.inventorySlots.size();
+        }
+
+        int xOffset = SlotLayoutManager.getXOffset(SlotLayoutManager.Mods.Galacticraft);
+
+        if (galacticraftAccessories == null) {
+            if (!player.worldObj.isRemote) {
+                EntityPlayerMP playerMP = PlayerUtil.getPlayerBaseServerFromPlayer(player, false);
+                this.galacticraftAccessories = GCPlayerStats.get(playerMP).extendedInventory;
+            } else {
+                this.galacticraftAccessories = ClientProxyCore.dummyInventory;
+            }
+        }
+
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 6, xOffset, 8 + 0 * 18, player, GC_THERMAL_HELM));
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 7, xOffset, 8 + 1 * 18, player, GC_THERMAL_CHEST));
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 8, xOffset, 8 + 2 * 18, player, GC_THERMAL_LEGS));
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 9, xOffset, 8 + 3 * 18, player, GC_THERMAL_BOOTS));
+        xOffset += 18;
+
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 4, xOffset, 8 + 0 * 18, player, GC_PARACHUTE));
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 0, xOffset, 8 + 1 * 18, player, GC_OXYGEN_MASK));
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 2, xOffset, 8 + 2 * 18, player, GC_OXYGEN_TANK));
+//        nullSlots.add(Pair.of(xOffset, 8 + 3 * 18));
+        xOffset += 18;
+
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 5, xOffset, 8 + 0 * 18, player, GC_FREQUENCY_MODULE));
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 1, xOffset, 8 + 1 * 18, player, GC_OXYGEN_GEAR));
+        this.addSlotToContainer(new SlotDWS(galacticraftAccessories, 3, xOffset, 8 + 2 * 18, player, GC_OXYGEN_TANK));
+//        nullSlots.add(Pair.of(xOffset, 8 + 3 * 18));
+    }
+}
